emacs: add an option for the shell launching Emacs
This patch adds an option services.emacs.shell to choose the shell
launching an Emacs daemon.

This option defaults to bash (runtimeShell), which has backword
compatibility, and should be set to the user's shell to get the right
environment for Emacs.

Before, this shell is hardcoded to runtimeShell, i.e. bash. If a
user's shell is not bash and programs.bash.enable is false, the Emacs
daemon will miss some environment variables,
e.g. home.sessionVariables and programs.zsh.sessionVariables (if zsh
is the user's shell).

Fixes #2139